Create a `Bot` monad to encapsulate functions
This introduces a `Bot` monad that is a combination of the `Reader` and
`State` monads. The `Reader` lets us hold onto App configuration data,
like socket connection (can be a more complex data structure later). The
`State` monad is currently just counting the activity happening, but we
can leverage it to store message data for processing.