First steps toward Freedom
Experiment with the Free monad. Create 2 DSLs, one for the main app
commands and one for the message specific commands. This uses a second
mtl combination for single `ChannelState` operations. The second custom
monad lets us abstract the `ChannelState` from the inputs as well as do
what we need to do with it and get back the final version then set that
back to the `Bot` state.

The Free monad lets us abstract a DSL of the main commands out of the
code. Once we make the DSL a `Functor` then using the Free monad we
combine it to get a monad for free. Then we can create an interpretor to
do what we need when it is time.