Fix Heroku deployments
Merely pushing to the registry doesn't deploy to Heroku. The additional
container:release step is required.

Thoughtbot's Heroku app name was hardcoded in the script. The new
`HEROKU_APP_NAME` env variable allows it to be changed per installation.

Running as a "web" process failed because Heroku expected it to listen
to a port. This reclassifies it as a "worker" process.

REDIS_URL is added to `env` because it's a required env variable.